cleanup header dependencies
In their current state, the headers could not be analysed by clang-tidy
because there was (almost) no include.

The only way to make NTL works was to include ntl.h (which basically
include everything in a single file so it ends up working…).

That means that an header, taken independantly, cannot not be "compile"
(e.g dft.h use RN, but doesn't include gf.h, …).