use namespaces
We introduce the following namespaces:
- nttec: the root namespace of the NTTEC library
- nttec::arith: for the base/core arithmetical functions of NTTEC.
- nttec::fec: for the code related to Forward Error Correction.
- nttec::gf: for the code related to Galois Fields.
- nttec::fft: for the code related to the Fast Fourier Transform.
- nttec::vec: for the low-level wrappers around memory buffer.

Note that we add to move the overloading of << for the 128-bit integers
into the namespace "std", otherwise we have some problem to find it when
needed (ADL cannot find it in some cases).

Also move pack/unpack helper of Vec from `fec.h` to `vec.h`: it's more
logical that way.

Closes: #155