export the version string
The version string is available in VERSION, but this define is mainly
for internal purpose (it will probably be overwritten when used
externally, since it has such a common name…).

This commit adds a function to export the version string to the
users/clients of the library.

Also, don't expose `build_info.h`: this is a private header.