improve dependencies handling
By using target-specific instructions we can avoid duplication and have
a better (more robust) dependencies management.

See https://floooh.github.io/2016/01/12/cmake-dependency-juggling.html

CMake can handle transitive dependencies, so we only have to set the
include directories and link libraries for the libraries (even though
they don't link against anything).

Note that link libraries cannot be specified for OBJECT libraries,
that's why we have to use a for loop to set them for both the STATIC and
SHARED libraries.
Another limitation of OBJECT libraries is that they don't "forward"
their dependencies, so we have to manually forward them (for both
compile definitions and include directories).

For compile definition we can simply copy the property's value, but for
the include directories we have to do it "manually" in order to keep the
distinction between SYSTEM and "local" includes.