add installation targets and installers
NTTEC is composed of two parts:
- the libraries (static and shared): required to develop/run softwares
using NTTEC
- the C++ headers: required to develop softwares using NTTEC

The GNUInstallDirs module is used to deal with system-specific
peculiatities regarding standard locations (lib vs lib64, …).

Use the CPack module to generate binary and source installers.
In this commit we only generate tarballs, but in later commit we may
introduce support for DEB or RPM packages (and Apple-stuff as well
maybe).

Note that we take care to create components (and components groups) that
can be used for binary installers that support it.
See https://cmake.org/Wiki/CMake:Component_Install_With_CPack