add "bugprone" and "cert" lint checks
We disable `cert-dcl58-cpp` because it doesn't seems to make the
distinction between valid and invalid modification of the `std`
namespace.

Thanks to `cert-dcl03-c`, a misuse of assert was detected:
`assert("a string")` is always true, and won't show an error message).

Also:
- we replace `atoi` by `std::stoi` (cert-err34-c)
- exception handling: throw by value, catch reference (cert-err61-cpp
  and cert-err09-cpp)
- ignore cert-err58-cpp for the benchmark tool
- replace `rand` by C++11 random library (cert-msc50-cpp and
  cert-msc30-c)¹

Also, by replacing `rand` I noticed that there was a bug in prng.h
(calling `rand` instead of `_rand`).

¹: Note that even though the `weak_rand*` functions were generic on
   their return type, they only provided 16 to 32-bit of entropy
   (because they were directly returning the return value of `rand` wich
   is an int (which is usually 32-bit)).

   In order to be "compatible", I used a 32-bit Mersenne-Twister and a
   uniform distribution on 32-bit. We could upgrade to 64-bit if we want
   (but then, I notice that some tests were stuck/very slow (like
   `test_factor_distinct_prime`, probably because the algorithm has
   trouble with "big" 64-bit numbers)).