RingModN: const member functions & primitive root calculation
- Declare const member functions
- Primitive root calculation is performed by default in the constructor
of RingModN. However, this operation uses operations such as mul, exp
etc. that depends on the target finite field. Therefore, this operation
should be executed only in the constructor of the target class.
For example with gf::Extension extending from gf::Field that extends
from gf::RingModN, this calculation should be executed in the
constructor of gf::Extension. Hence, we need a variable to control it.
That is the 3rd argument indicate whether the computing of primitive
root is performed in the constructor of RingModN.