don't call virtual functions during construction
It seems that simplifying the code helped Clang analyzer to find an
issue (yet another reason to have simple code).

Consider the following callstacks:
- Call `RingModN` constructor with `calculate_root=true`
- Call `RingModN<T>::init` (because `calculate_root=true`)
- Call `RingModN<T>::find_primitive_root`
- Call `RingModN<T>::exp` from inside `RingModN<T>::find_primitive_root`
- Call `RingModN<T>::exp_quick`
- Call `this->mul(tmp, tmp);`

The last call is a call to a virtual function and we are in the
construction: this can be tricky (see[1]).

To solve this issue, we will use a two phase initialization:
- first we create the object
- then call the init method

To avoid any misuse (forget to call `init`) we create a "factory
function" that is responsible of the creation of all the GF objects (and
to force the user to use the factory function, we make the constructors
private).

Note that because `gf::Field` and `gf::NF4` contain an unique_ptr, they
cannot be copied so we need to define a move constructor for them.

[1]: https://wiki.sei.cmu.edu/confluence/display/cplusplus/OOP50-CPP.+Do+not+invoke+virtual+functions+from+constructors+or+destructors