FFT2n: enhance iFFT
Previously, iFFT does two things on input:
- shuffle output's elements according to an reversed bit order.
- modify input vector/buffer
This change will fix these things.

Note, Input buffer is in reversed-bit order. Hence butterfly operations
can be performed directly on this order. But the output must be re-order
in the natural order by using swap operations.
For FFT on buffers, a swap operation is to swap only two buffers'
pointer, hence it's not costly. However, it will modify the output
buffers. To avoid that, we shuffle the output twice, before and after
the butterfly operation.