fix TestFftAdd
When `weak_rand` returns 2, the test fails with a FPE because we divide
by zero when calculating `t` in `test_taylor_expand`.

Note that fixing this is not enough, because the Taylor Expension expect
m >= 2 so anyway, we should execute theses test with `len = 2` (this is
properly handled, when calling `fft` but not when calling directly
`taylor_expand`.

Fix this by removing the call to `weak_rand` (it mades the test
non-reproducible anyway…).

Also fix an issue on the computation of `t` and `n` to ensure that 1 < t <= n