fix `-Woverloaded-virtual`
Here is a quick drill down, for those unfamiliar with this aspect of
C++.

What's is this about?
=====================

The first thing to note is that there is no overloading across scopes,
and derived class scopes are not an exception.

Concretely, that means that this code:

    struct A {
        int foo(int i) { std::cout << "foo(int): "; return i+1; }
    };

    struct B : public A {
        double f(double d) { std::cout << "foo(double): "; return d+1.3; }
    };
    […]
    B* pd = new B;
    std::cout << pb->f(2) << '\n';

Will produce:

    f(double): 3.3

Because the best method is only looked up into D scope, so even though a
better match is available for `int` in B, it's not found and the version
for `double` is used.

The compiler warning is triggered when the hidden function is virtual,
as shown in the example from the man:

    struct A {
        virtual void f();
    };

    struct B: public A {
        void f(int);
    };

the A class version of f is hidden in B, and code like:

    B* b;
    b->f();

fails to compile.

What about us?
==============

Our codebase is affected by this problem, the hidden functions are:
- void FecCode<T>::decode_apply(const DecodeContext<T>&, Buffers<T>&, Buffers<T>&)
- void FecCode<T>::decode_prepare(const DecodeContext<T>&, const std::vector<Properties>&, off_t, Buffers<T>&)
- void FecCode<T>::encode_post_process(Buffers<T>&, std::vector<Properties>&, off_t)
- void FecCode<T>::encode(Buffers<T>&, std::vector<quadiron::Properties>&, off_t, Buffers<T>&)
- void FourierTransform<T>::fft_inv(Buffers<T>&, Buffers<T>&)
- void FourierTransform<T>::fft(Buffers<T>&, Buffers<T>&)
- void FourierTransform<T>::ifft(Buffers<T>&, Buffers<T>&)
- void RingModN<T>::neg(Buffers<T>&) const
- void RingModN<T>::neg(size_t, T*) const

You can see that most of them are `Buffers` version of pure abtract
methods that works on `Vector`.

Since the `Vector` versions are pure abstract, children classes must
override them and by doing so mask the overloaded versions that work on
`Buffers` (except if they also override the `Buffers` versions`, but
that's not always the case).

Example:
- `fft::Radix2` override both `Vector` and `Buffers` versions
  => no compiler warning.
- `fft::CooleyTukey` doesn't override the `Buffers` versions
  => we have a warning.

But our code compile/works!
===========================

Yeah, by happenstance.

Because most of these `Buffers` version are defined as no-op by default
and if a child class doesn't override it that propably means that it's
never called and thus yeah, our code compile.

But that doesn't change the fact that the code is brittle, and we should
clean these up!

How do we fix this?
===================

By explicitely asking for the methods lookup to consider the scope of
the parent class, which is done with an using-declaration.

Appendix: why this weird and confusing behavior‽
================================================

Because if we were looking up into the parents' scopes, then the code
have certain unforeseen and potentially dangerous behavior.

Consider the following code:

    class Base {
        int x;
    public:
        virtual void copy(Base* p) { x = p-> x; }
    };

    class Derived{
        int xx;
    public:
        virtual void copy(Derived* p) { xx = p->xx; Base::copy(p); }
    };

    void f(Base a, Derived b)
    {
        a.copy(&b); // ok: copy Base part of b
        b.copy(&a); // BUG: Base::copy is called and `xx` is not copied!
    }

If we were looking up into the parent's scope, then we would find
`Base::copy(Base*)` which is a better match for the second call to copy
BUT it would lead to object slicing, which is unexpected here!

References
==========

- http://www.stroustrup.com/bs_faq2.html#overloadderived
- https://stackoverflow.com/a/1629074