ci: test build in debug mode with all compiler & all SIMD mode
The debug mode is the one with the more warnings, so this is the one
that should be tested (if it compiles in Debug, it will compile in
Release), but since we use conditionnal compilation for SIMD we should
make sure that every codepath is clean.

Compilation in all SIMD mode in release is done by the test job anyway.