Property: use vector to store location & marker
We use std::vector to store pairs of location and marker. It gives
better performance compared to map or unordered_map for our operations:
addition & accessing.

Util functions are added:
- location(i): get location of the ith element
- marker(i): get marker of the ith element
- is_marked(i, loc): check if location of the ith element == loc
- in_range(i, min, max): check if location of the ith element is within
the range [min, max)

Note, in these util functions, we assume that stored elements are sorted
in ascending order of location.