RsNf4 works with updated in FecBase
In FecBase's encode and decode vertically, the steps of casting from
`char` data to `T` elements is removed due to the new `Buffers`.
However, `RsNf4` works currently with `Buffers` without meta. Hence, the
steps will be integrated in RsNf4's encode and decode.
We use an internal buffer in RsNf4 for the casting steps. Thanks to the
buffer, input source data of encoding will not be modified.